add transak to instant dapp package
